split up world deployment scripts
