+++ packages/client/src/layers/network/api/world/index.ts
@@ -1,0 +1,3 @@
+import { WorldAPI, setupWorldAPI } from './world';
+export { setupWorldAPI };
+export type { WorldAPI };
